+++ waspddi/DDINodeView.cpp
@@ -7,7 +7,6 @@ DDINodeView::DDINodeView(std::size_t                node_index,
                          const AbstractInterpreter& pool)
     : m_node_index(node_index), m_pool(&pool)
 {
-    wasp_require(!this->is_null());
 }
 
 DDINodeView::DDINodeView(const DDINodeView& orig)