+++ waspjson/JSONObjectParser.cpp
@@ -110,7 +110,8 @@
 
 
 // Suppress unused-variable warnings by "using" E.
-#define YYUSE(E) ((void) (E))
+//#define YYUSE(E) ((void)(E))
+
 
 // Enable debugging if requested.
 #if YYDEBUG