+++ waspson/SONNodeView.cpp
@@ -7,7 +7,6 @@ SONNodeView::SONNodeView(std::size_t                node_index,
                          const AbstractInterpreter& pool)
     : m_node_index(node_index), m_pool(&pool)
 {
-    wasp_require(!this->is_null());
 }
 
 SONNodeView::SONNodeView(const SONNodeView& orig)