+++ waspvii/VIINodeView.cpp
@@ -26,7 +26,6 @@ VIINodeView::VIINodeView(std::size_t                node_index,
                          const AbstractInterpreter& pool)
     : m_node_index(node_index), m_pool(&pool)
 {
-    wasp_require(!this->is_null());
 }
 
 VIINodeView::VIINodeView(const VIINodeView& orig)