+++ waspvii/VIIParser.cpp
@@ -106,7 +106,7 @@
 
 
 // Suppress unused-variable warnings by "using" E.
-#define YYUSE(E) ((void) (E))
+//#define YYUSE(E) ((void) (E))
 
 // Enable debugging if requested.
 #if YYDEBUG