+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
+<<<<<<< CHANGELOG.md
 - GCC 5.4 compiler compatibility CI job.
 - wasplsp - Language Server Protocol package with WaspServer and LSPInterpreter. See WASPLSP readme for details.
 - HIVE UNKNOWN input node validation capability. See HIVE readme for details.
+- waspdakota: a package for the Dakota UQ coupling. Specifically, support for the Dakota Parameter file format
+- wasputils/substitute: a utility that allows direct use of the Dakota parameter file for template attribute substitution
+
 
 ### Changed
 - wasvii hanging comment (comment 1 or more lines below last parameter) logic to reparent to the block